refactor: moved all type declarations into types file
